feat: cleanup and finalize

+++ include/Vec3.hpp
@@ -9,7 +9,7 @@ struct Vec3
     T z;
 };
 
-using Position = Vec3<int>;
+using Position = Vec3<float>;
 using Rotation = Vec3<int>;
 using Scale = Vec3<float>;
